make additional_ca array shapes match count matrix
`additional_ca` elements for `run10x()` are created from the CellRanger tSNE and Cluster files.  These may contain data for cells that are not present in the count matrix generated by Velocyto, causing loompy to fail to write due to different array shapes.  This adds a check that, if the shapes do not match, attempts to filter out the `additional_ca` elements that match discarded cells.